requirements.txt
fastapi
uvicorn
ollama
openai
python-multipart
python-dotenv
pillow
google-generativeai
mysql-connector-python